+++ frontend/app/components/forms/AppFormField.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFormikContext } from 'formik';
 
-import AppTextInput from './AppTextInput';
+import AppTextInput from '../AppTextInput';
 import ErrorMessage from './ErrorMessage';
 
 function AppFormField({ name, ...otherProps }) {
