+++ frontend/app/components/forms/ErrorMessage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
 
-import AppText from './AppText';
+import AppText from '../AppText';
 
 function ErrorMessage({ error, visible }) {
     if(!visible || !error) return null;
