+++ frontend/app/components/forms/SubmitButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFormikContext } from 'formik';
 
-import AppButton from './AppButton';
+import AppButton from '../AppButton';
 
 function SubmitButton({ title }) {
     const { handleSubmit } = useFormikContext();
