+++ frontend/app/screens/LoginScreen.js
@@ -1,15 +1,9 @@
 import React from 'react';
-import { StyleSheet, Image, TouchableNativeFeedbackBase } from 'react-native';
-import { Formik } from 'formik';
+import { StyleSheet, Image } from 'react-native';
 import * as Yup from 'yup';
 
-import AppButton from '../components/AppButton';
-import AppTextInput from '../components/AppTextInput';
-import ErrorMessage from '../components/ErrorMessage';
 import Screen from '../components/Screen';
-import AppFormField from '../components/AppFormField';
-import SubmitButton from '../components/SubmitButton';
-import AppForm from '../components/AppForm';
+import {AppForm, AppFormField, SubmitButton} from '../components/forms';
 
 const validationSchema = Yup.object().shape({
     username: Yup.string().required().label("Username"),
